--- backend/app/api/v1/auth.py
@@ -1,1 +1,0 @@
-
