--- backend/app/api/v1/email.py
@@ -1,1 +1,0 @@
-
