--- backend/app/api/v1/user.py
@@ -1,1 +1,0 @@
-
