--- backend/app/core/config.py
@@ -1,1 +1,0 @@
-
