--- backend/app/core/security.py
@@ -1,1 +1,0 @@
-
