+++ config/.env.example
@@ -1,1 +1,7 @@
- 
+EMAIL_ACCOUNT=####@gmail.com
+
+GOOGLE_CLIENT_ID=####.apps.googleusercontent.com
+GOOGLE_CLIENT_SECRET=####GOCSPX-
+
+OPENAI_API_KEY=####sk-proj-
+
