--- proto/requirements.txt
@@ -1,7 +1,0 @@
-fastapi==0.115.6
-uvicorn==0.32.1
-google-api-python-client==2.154.0
-google-auth-oauthlib==1.2.1
-IMAPClient==3.0.1
-openai==1.56.1
-python-dotenv==1.0.1
